fix(admin): much more nice code style
* fixed column text-align in table
