+++ src/modules/users.js
@@ -1,7 +1,8 @@
+import { toast } from 'react-toastify';
 import { API } from '../utils/api';
-import { SET_VISIBLE, SET_SEARCH_USER } from './userEntry';
 
 export const SET_USERS = 'users/SET_USERS';
+export const SET_LOOKUP_USER = 'users/LOOKUP_USER';
 
 const initialState = {
   isFetched: false,
@@ -10,14 +11,15 @@ const initialState = {
   page: 0,
   itemsPerPage: 25,
   filters: {},
+  lookupUser: null,
 };
 
 const format = (users) => {
   return users.map((user) => ({
     ...user,
     fullname: `${user.firstname} ${user.lastname}`,
-    tournamentName: user.team ? user.team.tournamentId : '',
-    teamName: user.team ? user.team.name : user.type === 'visitor' ? '(spectateur)' : '',
+    tournamentName: user.team ? user.team.tournament.name : '',
+    teamName: user.team ? user.team.name : user.type === 'spectator' ? '(spectateur)' : '',
     paidLabel: user.hasPaid ? '✔' : '✖',
     scannedLabel: user.scanned ? '✔' : '✖',
     permissionsLabel: user.permissions.join(', ') || '',
@@ -30,6 +32,11 @@ export default (state = initialState, action) => {
       return {
         ...state,
         ...action,
+      };
+    case SET_LOOKUP_USER:
+      return {
+        ...state,
+        lookupUser: action.lookupUser,
       };
     default:
       return state;
@@ -65,14 +72,28 @@ export const fetchUsers =
     });
   };
 
-export const displayUser = (user) => async (dispatch) => {
+export const lookupUser = (user) => async (dispatch) => {
+  const res = user ? await API.get(`admin/users/${user.id}/carts`) : null;
   dispatch({
-    type: SET_SEARCH_USER,
-    searchUser: user,
-  });
-  dispatch({
-    type: SET_VISIBLE,
-    visible: true,
+    type: SET_LOOKUP_USER,
+    lookupUser: user
+      ? {
+          id: user.id,
+          lastname: user.lastname,
+          firstname: user.firstname,
+          username: user.username,
+          email: user.email,
+          type: user.type,
+          age: user.age,
+          permissions: user.permissions,
+          hasPaid: user.hasPaid,
+          place: user.place,
+          discordId: user.discordId,
+          team: user.team,
+          attendant: user.attendant,
+          carts: res.data,
+        }
+      : null,
   });
 };
 
@@ -84,4 +105,27 @@ export const updateUser = (updateUser) => async (dispatch, getState) => {
     type: SET_USERS,
     users: formatUsers,
   });
+};
+
+export const validatePay = (id) => async (dispatch, getState) => {
+  const userModal = getState().users.lookupUser;
+  await API.post(`admin/users/${id}/force-pay`);
+  toast.success('Paiement validé');
+  dispatch(updateUser({ ...userModal, hasPaid: true }));
+  dispatch(lookupUser({ ...userModal, hasPaid: true }));
+};
+
+export const saveUser = (id, body, username) => async (dispatch, getState) => {
+  const userModal = getState().users.lookupUser;
+  const { data: user } = await API.patch(`admin/users/${id}`, body);
+  toast.success(`${username} mis à jour`);
+  dispatch(updateUser({ ...userModal, ...user }));
+};
+
+export const refundCart = (id) => async (dispatch, getState) => {
+  await API.post(`admin/carts/${id}/refund`);
+  const userModal = getState().users.lookupUser;
+  toast.success('Le panier a été marqué comme remboursé');
+  dispatch(updateUser({ ...userModal, hasPaid: false }));
+  dispatch(lookupUser({ ...userModal, hasPaid: false }));
 };
